+++ fheroes2/src/fheroes2/battle/battle_bridge.h
@@ -34,16 +34,16 @@ namespace Battle
     public:
 	Bridge();
 
-	bool NeedAction(const Unit &, s16) const;
-	void Action(const Unit &, s16);
+	bool NeedAction(const Unit &, s32) const;
+	void Action(const Unit &, s32);
 
 	void SetDestroy(void);
 	void SetDown(bool);
 	void SetPassable(const Unit &);
 
 	bool AllowUp(void) const;
-	bool NeedDown(const Unit &, s16) const;
-	bool isPassable(u8) const;
+	bool NeedDown(const Unit &, s32) const;
+	bool isPassable(int) const;
 	bool isValid(void) const;
 	bool isDestroy(void) const;
 	bool isDown(void) const;
