+++ fheroes2/src/fheroes2/game/difficulty.cpp
@@ -22,7 +22,7 @@
 
 #include "difficulty.h"
 
-const char* Difficulty::String(u8 difficulty)
+const char* Difficulty::String(int difficulty)
 {
     const char* str_difficulty[] = { _("difficulty|Easy"), _("difficulty|Normal"), _("difficulty|Hard"), _("difficulty|Expert"), _("difficulty|Impossible"), "Unknown" };
