+++ fheroes2/src/fheroes2/gui/statusbar.cpp
@@ -33,7 +33,7 @@ void StatusBar::SetCenter(const Point & pt)
     center = pt;
 }
 
-void StatusBar::SetCenter(s16 cx, s16 cy)
+void StatusBar::SetCenter(s32 cx, s32 cy)
 {
     center.x = cx;
     center.y = cy;
